requirements.txt
fastapi>=0.115.12
joblib>=1.4.0
joblib>=1.5.1
numpy>=2.2.6
pandas>=2.3.0
plotly>=6.1.2
pydantic>=2.11.5
python-dotenv>=1.1.0
Requests>=2.32.3
streamlit>=1.45.1
streamlit_extras>=0.7.1
uvicorn>=0.34.3
catboost>=1.2.8
scikit-learn>=1.6.1
scipy~=1.15.3
xgboost~=3.0.1
flaml~=2.3.4